Revert "Feat/emma"
