--- .github/workflows/pipeline.yml
@@ -1,8 +1,0 @@
-on: push
-
-jobs:
-  hello:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Echo Hello
-        run: echo 'Hello Emma this is a simple pipeline'
