+++ pom.xml
@@ -35,14 +35,9 @@
 			<optional>true</optional>
 		</dependency>
 		<dependency>
-<<<<<<< HEAD
 			<groupId>com.h2database</groupId>
 			<artifactId>h2</artifactId>
 			<scope>runtime</scope>
-=======
-			<groupId>org.postgresql</groupId>
-			<artifactId>postgresql</artifactId>
->>>>>>> cd91635e21845a087d587d774ba0d473e0b1fc00
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
