+++ src/main/resources/application.properties
@@ -1,4 +1,1 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
-spring.datasource.username=postgres
-spring.datasource.password=P0ok1e66
