+++ src/test/java/com/evergreen/evergreen/EvergreenApplicationTests.java
@@ -3,11 +3,11 @@ package com.evergreen.evergreen;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-// @SpringBootTest
-// class EvergreenApplicationTests {
+@SpringBootTest
+class EvergreenApplicationTests {
 
-// 	@Test
-// 	void contextLoads() {
-// 	}
+	@Test
+	void contextLoads() {
+	}
 
-// }
+}
