+++ src/test/resources/application.properties
@@ -1,5 +1,1 @@
 
-spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
-spring.datasource.username=postgres
-spring.datasource.password=password
-
