+++ src/test/resources/data.sql
@@ -1,5 +1,3 @@
-
--- INSERT INTO plant (id, name, size, species, age, status) VALUES
--- (1, 'Name 1', 50, 'Species 1', 5, 'Alive'),
--- (2, 'Name 2', 150, 'Species 2', 14, 'Alive');
-
+INSERT INTO plant (id, name, size, species, age, status) VALUES
+(1, 'Name 1', 50, 'Species 1', 5, 'Alive'),
+(2, 'Name 2', 150, 'Species 2', 14, 'Alive');
